added some infrastructure types, changed values serialization =============================================================
System Information:
Engine Version: Atom Ver 0.22.2.20200207
Library Version:  Ver 0.18.9.20200207

OS/HW Interface:
Windows Ver 0.11.2.20200205
Linux Ver 0.11.1.20200207

Hosts:
RX Interactive Console Ver 0.15.1.20200207
RX Pipe Child Ver 0.4.3.20200207
RX Simple Host Ver 0.2.0.20200207

Storages:
File System Storage Ver 0.3.4.20200207

Compilers:
MSVC 14.24.28316
GCC 8.3.0

OSes:
Linux Ver:4.15.0-76-generic [x86_64]
Windows Workstation 10.0.18362 [x64]

+++ lib/rx_io.h
@@ -1098,7 +1098,7 @@ stream_chuks_decoder<headerT,bufferT>::stream_chuks_decoder (std::function<bool(
         collected_(0),
         expected_(0)
 {
-	temp_byte_header_ = (uint32_t*)&temp_header_;
+	temp_byte_header_ = (uint8_t*)&temp_header_;
 }
 
 
@@ -1106,7 +1106,7 @@ stream_chuks_decoder<headerT,bufferT>::stream_chuks_decoder (std::function<bool(
 template <class headerT, class bufferT>
 bool stream_chuks_decoder<headerT,bufferT>::push_bytes (const void* data, size_t count)
 {
-	uint32_t* buffer = (uint32_t*)data;
+	uint8_t* buffer = (uint8_t*)data;
 	if (header_ == NULL)
 	{
 		int head_size = (collected_header_ + (int)count < (int)sizeof(headerT)) ? (int)count : (int)sizeof(headerT) - collected_header_;
